# Updated deque to cache "len(d.head.v)-1"